+++ changelog.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0] - 2019-05-08
+### Added
+- Mouse trails examples: [Line, SemiCircle]
+### Fixed
+- Mouse trail lost y position when page scroll
+
 ## [0.0.0] - 2019-04-29
 ### Changed
 - Work in progress on Navbar that appear on scroll